requirements.txt
coverage==7.3.2
google-cloud-storage==2.13.0
boto3==1.28.76
pytest==7.4.3